api/app/requirements.txt
# FastAPI and Uvicorn for API
fastapi==0.95.1
uvicorn==0.22.0

# OCR service integration
requests==2.31.0
python-multipart

# Optional: For managing Python environment if necessary
python-dotenv==1.0.0
